add leaderboard children routes and implement the function in domains and leaderboard header
