+++ explorer/src/components/layout/LeaderboardHeader.tsx
@@ -1,5 +1,6 @@
 'use client'
 
+import { getSupportedHeaderLinks } from '@/utils/route'
 import { Bars3BottomRightIcon, MoonIcon, SunIcon } from '@heroicons/react/24/outline'
 import { LogoIcon } from 'components/icons'
 import { INTERNAL_ROUTES, Routes } from 'constants/routes'
@@ -19,27 +20,7 @@ export const LeaderboardHeader: FC = () => {
   const [isOpen, setIsOpen] = useState(false)
   const { network } = useIndexers()
 
-  const menuList = useMemo(
-    () => [
-      {
-        title: 'Top Farmers',
-        link: `/${network}/${Routes.leaderboard}/${INTERNAL_ROUTES.leaderboard.farmers}`,
-      },
-      {
-        title: 'Top Accounts',
-        link: `/${network}/${Routes.leaderboard}/${INTERNAL_ROUTES.leaderboard.accounts}`,
-      },
-      {
-        title: 'Top Operators',
-        link: `/${network}/${Routes.leaderboard}/${INTERNAL_ROUTES.leaderboard.operators}`,
-      },
-      {
-        title: 'Top Nominators',
-        link: `/${network}/${Routes.leaderboard}/${INTERNAL_ROUTES.leaderboard.nominators}`,
-      },
-    ],
-    [network],
-  )
+  const menuList = useMemo(() => getSupportedHeaderLinks(network, Routes.leaderboard), [network])
 
   return (
     <header className="body-font z-9 py-[30px] font-['Montserrat'] text-gray-600">
